deprecated/requirements.txt
flask-login==0.4.0
flask-script==2.0.5
mysql-python==1.2.5
sqlalchemy==1.1.6
enum==0.4.6
psutil==3.3.0
